pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "coderev"
version = "0.3.0"
description = "AI-powered code review CLI using Groq (Kimi K2) with GitHub Actions CI integration"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "CodeRev Team"}
]
keywords = ["code-review", "ai", "kimi", "groq", "cli", "git", "diff"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "groq>=0.9.0",
    "pydantic>=2.0.0",
    "typer>=0.12.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
coderev = "coderev.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["coderev"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
